added routes to webpage
