+++ package.json
@@ -15,6 +15,7 @@
     "react-bootstrap": "^2.0.0-rc.0",
     "react-dom": "^17.0.2",
     "react-redux": "^7.2.4",
+    "react-router-dom": "^5.3.0",
     "react-scripts": "4.0.3"
   },
   "scripts": {
