+++ src/Routes/AllRecipes.js
@@ -1,0 +1,10 @@
+
+// Shows all recipes in data base
+export default function AllRecipes(){
+
+    return(
+        <div>
+            <h1>ALL RECIPES ROUTE</h1>
+        </div>
+    );
+}
