+++ src/app/App.js
@@ -1,44 +1,29 @@
-import {React, useState } from 'react';
-import recipes from '../recipes.js';
+import { 
+  BrowserRouter as Router,
+  Switch,
+  Route,
+  Link
+} from 'react-router-dom';
+import RandomRecipe from '../Routes/RandomRecipe.js';
+import AllRecipes from '../Routes/AllRecipes.js';
 import Navbar from '../components/Navbar/Navbar.js';
-import Button from '../components/Button/Button.js';
-import RecipeImage from '../components/RecipeImage/RecipeImage.js';
-import FoodContentContainer from '../components/FoodContentContainer/FoodContentContainer.js'
 import './App.css';
 
 
 function App() {
-
-  const[recipe, setRecipe] = useState({})
-
-  // gets random recipe and updates the recipe state
-  const randomizeRecipe = () =>{
-
-    const recipeIndex = Math.floor(Math.random() * recipes.length); 
-
-    // prevents recipe from showing up twice in a row
-    if(recipe === recipes[recipeIndex]){
-      randomizeRecipe();
-      return;
-    };
-
-    setRecipe(recipes[recipeIndex])
-    
-  }
-
-  // display a recipe on load
-  window.onload = () => {
-    randomizeRecipe();
-  }
-
-
   return (
     <div className="App">
-      <Navbar />
-      <RecipeImage recipe={recipe}/>
-      <br/>
-      <Button randomizeRecipe={randomizeRecipe}/>
-      <FoodContentContainer recipe={recipe} />
+      <Router>
+        <Navbar />
+        <Switch>
+        <Route exact path='/'>
+         <RandomRecipe />
+        </Route>
+        <Route  path='/allrecipes'>
+          <AllRecipes />
+        </Route>
+        </Switch>  
+      </Router>
     </div>
   );
 }
